requirements.txt
python-dotenv==0.17.1
Flask==1.1.2